backend/pyproject.toml
[project]
name = "agentic-rag-backend"
version = "0.1.0"
description = "Backend for Agentic RAG pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "slowapi>=0.1.9",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.34",
    "python-multipart>=0.0.9",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.27.0",
    "tiktoken>=0.7.0",
    "langchain-text-splitters>=0.3.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "rank-bm25>=0.2.2",
    "groq>=0.9.0",
    "sse-starlette>=2.0.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "aiosqlite>=0.20.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-cohere>=0.4.0",
    "langchain-openai>=0.2.0",
    "tavily-python>=0.5.0",
    "numexpr>=2.10.0",
    "deepeval>=3.9.7",

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
